Fix wrong nl_inside_namespace after code_width wrapping
When eat_blanks_after_open_brace and/or eat_blanks_before_close_brace are true
and code wrapping occurs, nl_inside_namespace is not taken into account and
corresponding blank lines are eaten.

Adjust the behavior so that nl_inside_namespace is considered in such cases and
specified number of blank lines is enforced.

Fixes: #2209