Introducing absolute thresholds
By specifying a negative threshold the threshold will be treated
as absolute. This means that the item will never be shifted right
by more then the specified threshold, even if other longer items
are nearby.

This is introduced to prevent a few function proto types returning
complex types to shift all functions prototypes to the right hence
moving their parameter list to the far right.

Now it is easier to predict how far the items can be moved while
being aligned.

refs. #2254