Generate version at build time
Modify the logic used to generate the version information (which
includes the VCS SHA) to do so at build time, rather than at configure
time. This is much more reliable, as source changes will not necessarily
trigger a reconfigure.

Also, try to fix correct generation of the build/release flag for
multi-configuration generators. (This will probably require a partial
rebuild when switching configurations, but at least it should generate
the right version information, which is an improvement over the previous
situation. I'll attempt a better fix in a following commit.)