Add debug flag to version via preprocessor
Tweak how we generate the version information so that the debug flag is
handled by the preprocessor, rather than by make_version.py. This should
be slightly more reliable, since the flag's presence or absence is keyed
by the presence or absence of the preprocessor defining DEBUG, which
presumably is what also controls the features for which knowing the
build type is interesting. More importantly, this means that the textual
contents of uncrustify_version.h are not configuration dependent, so the
header will not change when switching build configurations. (This is
especially relevant for multi-configuration generators, as per note in
the previous commit.)