Add notion of non-fatal option warnings
Modify OptionWarning to introduce the notion of "minor" warnings. These
are warnings which will not cause uncrustify to exit with a non-success
value.

These will be used to implement deprecation warnings, and are important
for unit testing deprecated options (because otherwise the tests will
fail due to the deprecation warning).