Don't match braces of namespace alias
Tweak the code that tries to match namespace braces to recognize and
ignore namespace aliases, which don't have associated braces. Also,
check if a brace was found before trying to dereference the matching
token.

This fixes a crash if there is no brace anywhere in the input after a
namespace or namespace alias. The latter is perfectly valid code, and
while the former is malformed, it is better to complain than to crash.