Update our internal config with new lambda options
Update the config used to format our own sources to specify that we
don't need compatibility with pre-0.70 versions, and to enforce recently
added options for styling of C++11 lambdas.

For now, this doesn't affect anything (we have very few lambdas, and it
seems they all follow the new rules already), but this will help keep
things consistent going forward.