Improve sp_type_ellipsis
Modify how sp_type_ellipsis is applied to also apply it between an
indirection (CT_PTR_TYPE or CT_BYREF) and an ellipsis. This corrects
some formatting issues and is consistent with e.g. how sp_type_func is
applied.

Also, fix some issues in the test cases which were not valid C++ code
and were causing uncrustify to parse the code in a manner other than
what was intended. This should make the tests more robust. (Note that,
considering the manner in which the test code was wrong, uncrustify's
parsing is actually reasonable.)