Don't apply sp_before_square to lambdas
Fix how sp_before_square is applied to avoid incorrectly applying it to
lambdas. This can be seen in (arguably pathological) code where a
freestanding lambda appears as something other than the start of a line.
(This issue has likely gone unnoticed as it is incredibly unlikely to
show up in "real" code. It was noticed while testing some other
lambda-related issues in "test" code reduced to a bare minimum number of
characters.)