Improve class/struct parsing
Modify parsing so that `class` is treated more like `struct`. Add
PCF_INCOMPLETE flag used to mark class/struct tokens of a forward
declaration (as opposed to a definition). Use this to fix application of
nl_template_class_{decl,def}, which was previously keying on PCF_LVALUE,
which is neither correct nor reliable. Adjust some other spots that were
relying on the old parsing.

The new parsing is subtly different; in particular, the terminal
semicolon of a forward declaration is now tagged with the respective
parent type. This only seems to affect comment insertion. There is also
another issue with comment insertion for which the fix is interesting
enough to warrant a separate commit.