Add more fine-grained control for nl_template_*
Add a bunch of new options for when to add newlines following `template`
and its parameter list. This alters the prior functionality of
nl_template_class, which previously applied to anything, not just
classes.

Right now, there is a test failure due to a parsing bug that is newly
revealed by the modifications in how we apply newlines after a template
parameter list. This will be fixed in a subsequent commit. Additional
tests for the new options will also be added in a subsequent commit.

For classes and templates, we are providing very fine-grained control,
with the ability to override the base option differently for
declarations versus definitions, and to further override those options
for specializations. This is not yet implemented for variables (mostly
because it is less trivial to determine the difference between a
declaration and a definition), but the design is such that these
overrides can be added in the future with no change in behavior.