Don't mark struct name as PCF_IN_STRUCT
Remove the logic that marks everything from the `struct` keyword through
the opening brace as PCF_IN_STRUCT. Strictly speaking, we aren't *in*
the struct definition yet, and more importantly, this made it impossible
to distinguish a nested struct definition (i.e. one inside another
struct) from a top-level struct. While this apparently doesn't matter
for structs, it is critical for classes, which, since the previous
commit, are now subject to more consistent parsing. (In particular, this
was breaking cmt_insert_class_header.)