Don't put stack memory in ParseFrame
Fix a logic error in preproc_start, which was adding an entry to the
ParseFrame which pointed to stack memory (which immediately became
invalid). This fixes a subtle error where any subsequent access of this
entry would get garbage, because the token long since ceased to exist as
such, but was not showing up in memory checkers because the stack
address was still valid.

(This was discovered when unrelated changes perturbed things "just
right" so that the bit of stack being referenced "just happened" to be
flagged by ASAN, and thus finally triggered an error.)