Fix option compatibility in included configuration
Modify how we handle the option compatibility level to be more correctly
implemented when parsing nested configuration files via the `include`
directive. Previously, we had a global compatibility level, such that an
included configuration might change the compatibility level for the
configuration that included it. Now, the level is scoped to the file
being parsed.

Both behaviors are potentially useful, but this seems more likely to be
used in practice and less likely to surprise users.