Improve flag logging
Add a new helper function to convert pcf_flags_t to a human-readable
string. Use this in some places that were previously only logging the
raw value.

Besides being more human-readable, one of the changes avoids a bug in
GCC < 8.x where the enum type is accepted as variadic argument, but is
incorrectly converted to an unsigned long, which does not match the
format specifier (unsigned long long), resulting in a compiler warning.