Create a script to generate the CHANGELOG
Create a script that attempts to scan for added options and output a
list of such, along with their corresponding dates of addition. This is
somewhat brittle and can be confused by e.g. options being moved. I have
some ideas for how to improve it, but want to commit this initial
version for the sake of history.