Improve gen_changelog.py
Refactor how we try to parse changes to extract the complete "before"
and "after" option sets (fairly straightforward and robust) rather than
trying to parse patches (complicated and brittle). This makes the script
much more robust in general, and also allows us to report removed
options as well. (A removal typically happens in conjunction with an
addition, such that what actually happened is a rename, but trying to
sort that out in code is on the order of NLP-difficult, so that is left
to a human.)

Also tweak the date format to more closely match the actual changelog.
(In case it wasn't obvious, the purpose of this script is to drastically
simplify creation of the changelog, not to completely automate it.)