Tweak output of gen_changelog.py
Adjust the output of gen_changelog.py so that it is easier to convert to
the desired final format. I think this strikes a good balance between
minimal additional tweaking needed while still usefully providing the
information that helps a human sort out things like renamed or transient
options that need manual tweaking.