Fix formatting error in newlines.cpp
Fix a formatting inconsistency (actual code is different from how
uncrustify wants it formatted) in newlines.cpp that is causing the
sources-format test to fail. This probably happened due to the code
being modified in one branch that predates recent change to uncrustify's
formatting configuration which happened in parallel, resulting in a sort
of "soft conflict".