Create script to help automate releases
Create a new script to help perform several of the actions related to
creating a new release. In particular, this automates several file
changes that previously needed to be done by hand. (It isn't a "one
button" script, but it should at least cut down on the number of details
that need to be remembered.)

Documentation will be added/updated in a future commit.