Merge update-defaults.sh into release_tool.py
Consolidate the additional actions performed by update-defaults.sh into
the 'update' step of the new release tool. This is slightly more
portable, and there is no reason why the updates that update-defaults.sh
performed need to be a separate step. (Indeed, it makes more sense for
them to happen in the same commit as the other changes made by the
'update' step of the release tool.)