Teach release_tool.py to use pull requests
Modify release_tool.py to create a scratch branch while operating
(instead of a temporary tag that will be recreated later), so that this
branch can be pushed and merged as a pull request, rather than having
release_tool.py operate directly on the master branch. Also replace the
'push' command with a 'tag' command that just creates and pushes the
tag given an already-merged pull request.