Fix version in release_tool.py generated files
Since ceddebc3accb, we no longer create a "scratch" tag for preparing
releases, which means the uncrustify executable does not yet show the
new version number (as it did when we were using a "scratch" tag). This
means that the version number in files that get generated for the
release is not correct. Tweak release_tool.py to also munge the version
number in these files after generating them.