Fix double "Uncrustify" in version fallback
Rename CURRENT_VERSION variable to UNCRUSTIFY_VERSION, and remove the
name "Uncrustify" from its value. This fixes the fallback logic for
determining the version (when we are unable to determine the version
from the repository, as is the case when e.g. building from a tarball)
so that we don't end up with "Uncrustify-Uncrustify-X.Y.Z". This is also
generally less confusing and better practice, and allows us to also
excise the silly double name from the CPack packages and eliminate the
step to rename said packages from the release process.