Don't scare user if make_version.py fails
Change how GenerateVersionHeader.cmake reports failures of
make_version.py to look less scary. A failure is normal and expected
when building from a source package as opposed to a repository clone,
and the old reporting, even though it was non-fatal, was easy to mistake
as an actual problem. Hopefully the new wording will make it more
obvious that nothing is actually wrong.