Fix issue using NoGitVersionString option
When the CMake option NoGitVersionString is enabled, create a dummy
generate_version_header target that does nothing, so that when we later
try to add a dependency on this target (which, when NoGitVersionString
is OFF, would generate the version header), CMake doesn't complain about
there being no such target.

(n.b. When NoGitVersionString is ON, the header is generated at
configure time instead using the version hard-coded into our
CMakeLists.txt.)

(Yes, we could also just make the dependency conditional, but IMHO this
produces slightly cleaner CMake code for negligible cost.)