Bump minimum CMake version, fix warning
Increase the minimum required CMake version to 3.2. List
uncrustify_version.h as a BYPRODUCT of the rule that generates it, which
is relevant for generating a proper build graph for some tools (notably
Ninja) and fixes a CMake warning to such effect.

Note that support for BYPRODUCTS was introduced in CMake 3.2, which is
why these changes are being combined. (In a very real sense, fixing the
warning is the driver for increasing the minimum version, which in turn
is a more significant change than silencing a configure warning.)