fix preproc treated as type before func def
This fixes a bug where a preprocessor statement before a function
definition is treated as part of the type. This causes problems with
`#define` since chunks can be incorrectly modified.

This is fixed by using `chunk_get_next_ncnlnp()` to also skip over
preprocessor statements in addition to comments and newlines when
fixing up the type before the function definition.

The test cast catches this bug because the bug cause the `*` in
`sizeof(type) * (num)` to be changed from CT_ARITH to CT_PTR_TYPE which
causes the space after `*` to be deleted. When the bug is fixed, the
token type of the `*` is not modified and we get the expected result.

Fixes #2690