Silence VS "performance warning"
Visual studio has a really obnoxious "performance warning" that it likes
to spew in some instances when an integral type is implicitly converted
to a Boolean, even though there is nothing actually wrong with such
code. Silence an instance of this by adding a `!!` to make the
conversion explicit. (This should not affect the generated assembly, but
it gets rid of the stupid warning.)