Fix newlines before namespaces
This commit addresses an issue in which the can_increase_nl() function
fails to allow newlines to be added before a namespace, if say, the
preceding statment does not enforce a subsequent newline and the
closing brace of the namespace is the last token in the file. The
fixes imposed by this commit are further demonstrated by applying
the configuration specified by 'forUncrustifySources.cfg' to the
uncrustify source code itself. In 'uncrustify_limits.h', 'namespace
uncrustify' is preceded by a '#pragma once' statement, but there
should be a newline before the namespace according to the
configuration specified in 'forUncrustifySources.cfg'.