Fix functions treated as variables
Tweak code that detects variable definitions to skip over reference
types when looking for the identifier associated with a type (similar to
how the code was previously adjusted to do this in the face of pointer
types). This fixes a bug where a member function prototype in a class
definition which returns a reference type was incorrectly identified as
the start of a block of variable definitions, which resulted in
erroneous application of nl_var_def_blk_start.

Fixes #3018