Fix cmt_reflow_mode under code_width constraints
This commit fixes issues with comment reflow under code width
contraints (code_width > 0) for multi-line c-style comments,
i.e. /* ... */. With this fix, lines combined by the reflow and
then wrapped due to code width limitations will now have proper
indentation according to the sp_after_star_cont option. Note that
the indentation is applied *only* for the portion of the comment
that is wrapped to the next line. This fix applies under the
following conditions:

- cmt_indent_multi = true
- cmt_width > 0
- cmt_reflow_mode = 2

Added a unit test to demonstrate the fix.