Add ability to specify cmt reflow folding regex's
This commit introduces an option to specify a path to a file that contains
regular expressions describing patterns for which the end of one line and
the beginning of the next will be folded into the same sentence or paragraph
during full comment reflow. The regular expressions are described using
ECMAScript syntax. The syntax for this specification is as follows, where
"..." indicates the custom regular expression and "n" indicates the nth
end_of_prev_line_regex and beg_of_next_line_regex regular expression pair:

  end_of_prev_line_regex[1] = "...$"
  beg_of_next_line_regex[1] = "^..."
  end_of_prev_line_regex[2] = "...$"
  beg_of_next_line_regex[2] = "^..."
               .
               .
               .
  end_of_prev_line_regex[n] = "...$"
  beg_of_next_line_regex[n] = "^..."

  Note that use of this option overrides the default reflow fold regular
  expressions, which are internally defined as follows (to match legacy
  behavior of previously-hard coded non-regex criteria):

  end_of_prev_line_regex[1] = "[\w,\]\)]$"
  beg_of_next_line_regex[1] = "^[\w,\[\(]"
  end_of_prev_line_regex[2] = "\.$"
  beg_of_next_line_regex[2] = "^[A-Z]"