Add option to indent paragraphs in c-style commments during reflow
This commit adds an option to indent wrapped lines within c-style multi-
line comments (/* ... */) to the start of the current paragraph during
full comment reflow (cmt_reflow_mode = 2). A continuation indent is
calculated with respect to the start of the first sentence in a
paragraph, and this indent is applied to subsequent lines/sentences
that are part of the same paragraph. A separate continuation indent
is calculated for each subsequent/separate paragraph.