Properly fix nl_var_def_blk_start
Revert 947e149692e1 and instead tweak the code to not apply
nl_var_def_blk_start at the beginning of a block. This fixes a
regression introduced by the aforementioned commit, and also correctly
fixes the issue that commit was attempting to fix.

The unit test and several expected results have been updated
accordingly.