Fix another nl_var_def_blk_start bug
Modify the altered backward scanning from the previous commit that tries
to determine if a variable declaration immediately follows a brace (and
to use that, rather than the end of a statement, as the preceding
"thing" for the purpose of deciding whether to add a blank line) to skip
qualified members. It seems this may be what 947e149692e1 was actually
trying to accomplish.

It also seems like this should need to skip templates as well, but in
testing, introducing a template does not seem to cause malformatting.