1. Fix behavior of "chunk_get_next_ncnnlnp" and "chunk_get_prev_ncnnlnp"
From the name they are expected to return the next/prev chuck which is
not a comment, not a newline and not a preprocessor.
This was true only if the start chunk "cur" was a non preprocessor one.
Instead if "cur" was a preprocessor chunk, they would return either:
- the next non-preprocessor chuck
- the next preprocessor chuck that was both not a comment nor a new line.
The new version returns the expected chunk.

2. Add "chunk_get_next_ncnnl_in_pp" and "chunk_get_prev_ncnnl_in_pp"
functions, which return the next/prev preprocessor chuck that is
both not a comment nor a new line.

Signed-off-by: Michele Calgaro <michele.calgaro@yahoo.it>