Optimize include/import sorting speed
The current include/import sorting is a bit slow for files with a large
number of includes (100+). There are two primary factors causing this
slowness. The first is that the comparison is a bit expensive in
particular due to repeated regex matching. The second is the algorithm.
The current implementation makes use of a selection sort in order to
minimize the number of swaps but means a best case runtime performance
of *O(n)*.

This change implements memoization of the regex lookups to improve
comparison performance. This change also migrates the sorting
implementation to use `std::sort()` which offers best case runtime
performance of *O(n log n)*. Upon sorting values, the actual swaps are
performed so the worst case number of swaps is *O(n)*.