Make --check output more readable and relevant.
The current behavior of the --check flag is very suboptimal. It shows
the number of bytes that has changed or the current number of bytes,
which I don't believe is relevant for the user.

Additionally, it also shows which files have passed alongside those that
have failed. This is fine for a small project but becomes practically
impossible to read for projects with hundreds of files. Not using the -q
flag already tells us which files are being parsed, it's obvious files
that aren't shown to have failed have passed.