Fix trailing cmt alignment when braces move
If a newline is deleted before a brace, the brace may move to be before a comment
which was on the previous line. For example: 'do //comment\n{' may become
'do { //comment'. This breaks the existing logic for aligning comments.

After some discussion, this change fixes the issue by taking a broader approach than
a pinpoint fix for this specific case. First it makes Chunk::orig_prev_sp be
calculated correctly, and second it makes Chunk::orig_prev_sp be used as the
definitive source of "columns between this chunk and the previous one", instead of
the difference between current->orig_col and previous->orig_col_end.

There are still some places which might be appropriate to convert to use
orig_prev_sp, but that would be done in a future patch.