Makes mod_full_brace_if keep commented one liners
If "mod_full_brace_if = add" and an unbraced if() statement has a trailing comment,
then the added brace will be on a new line after the comment. However, this is true
even when "nl_if_leave_one_liners = true".

This change will keep the added brace on the same line as the if(), before the
comment, if it was a one-liner that was configured to be left.