Fix some else exprs with indent_ternary_operator==1
There are a number of valid expressions which do not get indented correctly when
indent_ternary_operator is set to 1. Rather than add many more if() tests to the
logic in indent_text() for these cases, I removed the tests for the ChunkType of the
chunk following a CT_COND_COLON. I don't think the type should matter when indenting
the 'else' branch, and this change passes the unit tests as well as on my codebase.