Fix parsing of a postfix incdec operator before a multiply
Without this fix, "result = *var++ * int(0x1111);" will have the second star parsed
as CT_DEREF instead of the correct CT_ARITH.

The only functional change here is the addition of "prev->Is(CT_INCDEC_AFTER)". The
remainder is just a reformatting of the existing code so it is much easier to read.