Adds a command to only update option document contents
This adds a command to release_tool.py which only updates the contents of the various
option documents. It does not alter any version information in them, nor does it run
any commands which contact remote git repos or alter the repo it runs in.

Since universalindent.cfg is another file that has to be updated when options change,
I added it to the new optiondocs command only, since I wasn't sure if you'd want it
during the normal update command.

This also updates CONTRIBUTING.md to show how and when to use it.

It also updates the regexes used to detect uncrustify version numbers to handle the
edge cases that come from not having a named tag or other baseline in git.