Adds C++ <cstdint> integer macros as types
C++11 defines a number of macros for minimum-width integer constants. This adds those
as types, so that uncrustify will "bind" them more tightly to their enclosed values
than it would for a function, which is how they get interpreted without this change.

For example, without this, uncrustify would be happy to break
"UINT64_C(0x0ab8e2a231b49732)" at the opening parenthesis, but with this change it
will not do that.

https://en.cppreference.com/w/cpp/types/integer#Function_macros_for_minimum-width_integer_constants