Fixed #3658 lambda indentation regression with namespaces
Removed namespace indent code introduced by my patch fix for issue #1813
The code has matured so that that fix is no longer needed, and actually
breaks things.

Secondly, due to misunderstanding of the indent_align_assign and
indent_continue settings, tests for issue #4328 were malformed. Updated
the configuration for those tests so that they are valid.