Make nl_var_def_blk_end work correctly w/classes
This is a smaller change than it seems. It reuses the logic for computing
the "prev" pointer from the variable-definition-block case for the class
(member) definition case. This way the code to add a newline to the end of
a variable definition block will do that instead of adding it to the end of
the "current" line, when that block is followed by class code.

The way it does this is to rearrange the code so that "prev" and "next"
pointers are computed for all cases inside the outermost if() clause, even
if they are not needed. This seems better than adding more redundant if()
clauses and having sometimes-uninitialized or sometimes-badly-initialized
pointers, or fully reorganizing the code.

This also fixes some logging statements in this block of code.